refactor(api): move non-entry modules into src/wrappers
Keep the src root reserved for package entry points; the handwritten SDK helpers now live under src/wrappers.
