+++ packages/api/src/index.ts
@@ -1,15 +1,19 @@
-export { awaitJobs } from './awaitJobs';
-export type { AwaitJobsOptions, AwaitJobsResult, JobResult } from './awaitJobs';
+export { awaitJobs } from './wrappers/awaitJobs';
+export type {
+  AwaitJobsOptions,
+  AwaitJobsResult,
+  JobResult,
+} from './wrappers/awaitJobs';
 export {
   decodeBase64,
   decodeFileContent,
   encodeBase64,
   encodeFileContent,
-} from './base64';
-export { DEFAULT_BATCH_SIZE, processBatches } from './batch';
-export type { BatchOptions } from './batch';
-export { API_VERSION, createApiClient } from './client';
-export type { ApiClientConfig, ApiVersion } from './client';
+} from './wrappers/base64';
+export { DEFAULT_BATCH_SIZE, processBatches } from './wrappers/batch';
+export type { BatchOptions } from './wrappers/batch';
+export { API_VERSION, createApiClient } from './wrappers/client';
+export type { ApiClientConfig, ApiVersion } from './wrappers/client';
 export * from './generated';
 export type { Client } from './generated/client';
-export type { RetryPolicy } from './transport';
+export type { RetryPolicy } from './wrappers/transport';
