+++ packages/api/src/wrappers/awaitJobs.ts
@@ -1,6 +1,6 @@
-import type { Client } from './generated/client';
-import { getTranslationJobInfo } from './generated/sdk.gen';
-import type { GetTranslationJobInfoResponse } from './generated/types.gen';
+import type { Client } from '../generated/client';
+import { getTranslationJobInfo } from '../generated/sdk.gen';
+import type { GetTranslationJobInfoResponse } from '../generated/types.gen';
 // Per-poll cap so a hung connection fails fast instead of eating the whole deadline.
 import { DEFAULT_TIMEOUT_MS as MAX_POLL_TIMEOUT_MS } from './transport';
